foundations_sdk/src/foundations/middleware/upstream_result_middleware.py
"""
Copyright (C) DeepLearning Financial Technologies Inc. - All Rights Reserved
Unauthorized copying, distribution, reproduction, publication, use of this file, via any medium is strictly prohibited
Proprietary and confidential
Written by Thomas Rogers <t.rogers@dessa.com>, 06 2018
"""

class UpstreamResultMiddleware(object):

    def call(self, upstream_result_callback, filler_builder, filler_kwargs, args, kwargs, callback):
        new_args = tuple(upstream_result_callback()) + tuple(args)

        return callback(new_args, kwargs)
